refactor: テストファイル構造を整理してVitestとPlaywrightを分離
- spec/unit/にユニットテストを移動（VitestとPlaywright混在問題解決）
- npm testでユニット+ブラウザテストの両方を実行するよう変更
- test:unitコマンドでVitestのみ実行可能
- .gitignoreにPlaywrightレポートディレクトリを追加
